requirements.txt
Django==4.2.7
google-generativeai==0.3.2
pdf2image==1.16.3
opencv-python==4.8.1.78
Pillow==10.1.0
spacy==3.7.2
python-magic==0.4.27
gunicorn==21.2.0
python-dotenv==1.0.0 